chore: Renaming node files and classes to be more uniform and differentiate between files and directories
