+++ src/yieldplotlib/load/exosims/__init__.py
@@ -1,7 +1,7 @@
 """Nodes for EXOSIMS specific data files."""
 
-__all__ = ["DRMNode", "EXOSIMSInputNode", "SPCNode"]
+__all__ = ["DRMFile", "EXOSIMSInputFile", "SPCFile"]
 
-from .drm_node import DRMNode
-from .exosims_input_node import EXOSIMSInputNode
-from .spc_node import SPCNode
+from .drm_file import DRMFile
+from .exosims_input_file import EXOSIMSInputFile
+from .spc_file import SPCFile
