+++ src/yieldplotlib/load/exosims/drm_file.py
@@ -1,9 +1,9 @@
 """Custom node for handling DRM pickle files."""
 
-from yieldplotlib.core.pickle_node import PickleNode
+from yieldplotlib.core.file_nodes import PickleFile
 
 
-class DRMNode(PickleNode):
+class DRMFile(PickleFile):
     """Node for handling DRM-specific pickle files."""
 
     def get(self, key: str):
